requirements.txt
matplotlib==3.3.4
mysql_connector_repackaged==0.3.1
pandas==1.2.4
SQLAlchemy==1.4.7
streamlit==1.1.0
